I Added functionality to the prev and next arrows, the button is an onClick button with a value of prevPerson and nextPerson.
I then created the  the next & prevPerson function at the top
, when the function is envoked it changes the state value of the index ,resulting in a
new rendered section . Remember that at the top we destructured
a particular set of infomation when we assigned the destructured property to
a people[index] value.

so as the Index changes when you click the button,it changes the values on the section.